rag_chatbot/requirements.txt
fastapi>=0.95.2
uvicorn>=0.18.3
langchain>=0.0.300
langchain-community>=0.0.6
ollama>=0.1.7
python-dotenv
numpy
pandas
sqlalchemy
psycopg2-binary
chromadb>=0.4.15
sentence-transformers
transformers
torch
redis
celery
prometheus-fastapi-instrumentator
pydantic[email]
passlib[bcrypt]
python-jose[cryptography]
aiohttp
tenacity
httpx